+++ src/main/java/com/calibermc/caliber/compat/block/CreateBlocks.java
@@ -1,8 +1,7 @@
 package com.calibermc.caliber.compat.block;
 
-import com.simibubi.create.AllBlocks;
+import com.calibermc.caliberlib.data.ModBlockFamily;
 import com.simibubi.create.content.decoration.palettes.*;
-import com.simibubi.create.foundation.data.CreateRegistrate;
 import com.calibermc.caliber.block.management.CaliberBlockManager;
 import com.calibermc.caliber.compat.item.CreateItems;
 import com.calibermc.caliberlib.block.management.BlockManager;
@@ -17,9 +16,7 @@ import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 import java.util.function.Supplier;
 
 import static com.calibermc.caliberlib.block.management.ModBlockHelper.*;
@@ -31,44 +28,58 @@ public class CreateBlocks {
 //    List&lt;Block&gt; block = ForgeRegistries.BLOCKS.getValues().stream().filter(value -&gt; ForgeRegistries.BLOCKS.getKey(value).getNamespace().equals("create")).toList();
     /* Stone */
 
-    public static final Block CUT_ASURINE;
-
+    public static final ArrayList<BlockManager> CREATE_MANAGERS = new ArrayList<>();
+    
     static {
-        CUT_ASURINE = ForgeRegistries.BLOCKS.getValue(new ResourceLocation("create", "cut_asurine"));
+        registerManagersFor("andesite", () -> BlockBehaviour.Properties.copy(Blocks.ANDESITE), AllPaletteStoneTypes.ANDESITE, STONE_VARIANTS_WITHOUT_SLAB_STAIRS_WALL);
+        registerManagersFor("asurine", () -> BlockBehaviour.Properties.copy(Blocks.DEEPSLATE), AllPaletteStoneTypes.ASURINE, STONE_VARIANTS);
+        registerManagersFor("calcite", () -> BlockBehaviour.Properties.copy(Blocks.CALCITE), AllPaletteStoneTypes.CALCITE, STONE_VARIANTS_WITHOUT_SLAB_STAIRS_WALL);
+        registerManagersFor("crimsite", () -> BlockBehaviour.Properties.copy(Blocks.DEEPSLATE), AllPaletteStoneTypes.CRIMSITE, STONE_VARIANTS);
+        registerManagersFor("deepslate", () -> BlockBehaviour.Properties.copy(Blocks.DEEPSLATE), AllPaletteStoneTypes.DEEPSLATE, STONE_VARIANTS_WITHOUT_SLAB_STAIRS_WALL);
+        registerManagersFor("diorite", () -> BlockBehaviour.Properties.copy(Blocks.DIORITE), AllPaletteStoneTypes.DIORITE, STONE_VARIANTS_WITHOUT_SLAB_STAIRS_WALL);
+        registerManagersFor("dripstone", () -> BlockBehaviour.Properties.copy(Blocks.DRIPSTONE_BLOCK), AllPaletteStoneTypes.DRIPSTONE, STONE_VARIANTS_WITHOUT_SLAB_STAIRS_WALL);
+        registerManagersFor("granite", () -> BlockBehaviour.Properties.copy(Blocks.GRANITE), AllPaletteStoneTypes.GRANITE, STONE_VARIANTS_WITHOUT_SLAB_STAIRS_WALL);
+        registerManagersFor("limestone", () -> BlockBehaviour.Properties.copy(Blocks.SANDSTONE), AllPaletteStoneTypes.LIMESTONE, STONE_VARIANTS);
+        registerManagersFor("ochrum", () -> BlockBehaviour.Properties.copy(Blocks.CALCITE), AllPaletteStoneTypes.OCHRUM, STONE_VARIANTS);
+        registerManagersFor("scoria", () -> BlockBehaviour.Properties.copy(Blocks.BLACKSTONE), AllPaletteStoneTypes.SCORIA, STONE_VARIANTS);
+        registerManagersFor("scorchia", () -> BlockBehaviour.Properties.copy(Blocks.BLACKSTONE), AllPaletteStoneTypes.SCORCHIA, STONE_VARIANTS);
+        registerManagersFor("tuff", () -> BlockBehaviour.Properties.copy(Blocks.TUFF), AllPaletteStoneTypes.TUFF, STONE_VARIANTS_WITHOUT_SLAB_STAIRS_WALL);
+        registerManagersFor("veridium", () -> BlockBehaviour.Properties.copy(Blocks.TUFF), AllPaletteStoneTypes.VERIDIUM, STONE_VARIANTS);
     }
 
-    /* Natural Stone */
-    public static final BlockManager CREATE_ASURINE = CaliberBlockManager.registerCreate("asurine_block", () -> BlockBehaviour.Properties.copy(Blocks.DEEPSLATE), () -> AllPaletteStoneTypes.ASURINE.baseBlock.get(), STONE_VARIANTS);
-    public static final BlockManager CREATE_CRIMSITE = CaliberBlockManager.registerCreate("crimsite_block", () -> BlockBehaviour.Properties.copy(Blocks.DEEPSLATE), () -> AllPaletteStoneTypes.CRIMSITE.baseBlock.get(), STONE_VARIANTS);
-    public static final BlockManager CREATE_LIMESTONE = CaliberBlockManager.registerCreate("limestone_block", () -> BlockBehaviour.Properties.copy(Blocks.SANDSTONE), () -> AllPaletteStoneTypes.LIMESTONE.baseBlock.get(), STONE_VARIANTS);
-    public static final BlockManager CREATE_OCHRUM = CaliberBlockManager.registerCreate("ochrum_block", () -> BlockBehaviour.Properties.copy(Blocks.CALCITE), () -> AllPaletteStoneTypes.OCHRUM.baseBlock.get(), STONE_VARIANTS);
-    public static final BlockManager CREATE_SCORIA = CaliberBlockManager.registerCreate("scoria_block", () -> BlockBehaviour.Properties.copy(Blocks.BLACKSTONE), () -> AllPaletteStoneTypes.SCORIA.baseBlock.get(), STONE_VARIANTS);
-    public static final BlockManager CREATE_SCORCHIA = CaliberBlockManager.registerCreate("scorchia_block", () -> BlockBehaviour.Properties.copy(Blocks.BLACKSTONE), () -> AllPaletteStoneTypes.SCORCHIA.baseBlock.get(), STONE_VARIANTS);
-    public static final BlockManager CREATE_VERIDIUM = CaliberBlockManager.registerCreate("veridium_block", () -> BlockBehaviour.Properties.copy(Blocks.TUFF), () -> AllPaletteStoneTypes.VERIDIUM.baseBlock.get(), STONE_VARIANTS);
+    public static void registerManagersFor(String name, Supplier<BlockBehaviour.Properties> properties, AllPaletteStoneTypes type, Collection<ModBlockFamily.Variant> variants) {
+        if (name.contains("asurine") || name.contains("crimsite") || name.contains("limestone") || name.contains("ochrum") || name.contains("scoria") || name.contains("scorchia") || name.contains("veridium")) {
+            CREATE_MANAGERS.add(CaliberBlockManager.registerCreate("%s_block".formatted(name), properties, () -> type.baseBlock.get(), variants));
+        }
 
-    /* Cut Stone */
-//    public static final BlockManager CREATE_CUT_ASURINE = CaliberBlockManager.registerCreate("cut_asurine", () -> BlockBehaviour.Properties.copy(Blocks.DEEPSLATE), () -> CUT_ASURINE, STONE_VARIANTS_WITHOUT_SLAB_STAIRS_WALL);
-
-
-//    public static final BlockManager CREATE_CUT_ASURINE = CaliberBlockManager.registerCreate("cut_asurine", () -> BlockBehaviour.Properties.copy(Blocks.DEEPSLATE), () -> Arrays.stream(AllPaletteStoneTypes.ASURINE.variantTypes).filter(v -> v.equals(PaletteBlockPartial.CUT)), STONE_VARIANTS_WITHOUT_SLAB_STAIRS_WALL);
-
-
-//    public static final BlockManager CREATE_CUT_ASURINE = CaliberBlockManager.registerCreate("cut_asurine", () -> BlockBehaviour.Properties.copy(Blocks.DEEPSLATE), () -> Arrays.stream(AllPaletteStoneTypes.ASURINE.variantTypes)
-//            .filter (v -> v.getTexture(paletteStoneVariants.name("cut_asurine")).get()), STONE_VARIANTS_WITHOUT_SLAB_STAIRS_WALL);
-
-
-    /* Cut Bricks */
-
-    /* Layered Stone */
-
-    /* Pillar Stone */
-
-    /* Polished Cut Stone */
-
-    /* Small Bricks */
+        for (PaletteBlockPattern variantType : type.variantTypes) {
+            String s = name;
+            if (variantType == PaletteBlockPattern.CUT) {
+                variants = STONE_VARIANTS_WITHOUT_SLAB_STAIRS_WALL;
+                s = "cut_" + s;
+            } else if (variantType == PaletteBlockPattern.BRICKS) {
+                variants = STONE_VARIANTS_WITHOUT_SLAB_STAIRS_WALL;
+                s = "cut_" + s + "_bricks";
+            } else if (variantType == PaletteBlockPattern.SMALL_BRICKS) {
+                variants = STONE_VARIANTS_WITHOUT_SLAB_STAIRS_WALL;
+                s = "small_" + s + "_bricks";
+            } else if (variantType == PaletteBlockPattern.POLISHED) {
+                variants = STONE_VARIANTS_WITHOUT_SLAB_STAIRS_WALL;
+                s = "polished_cut_" + s;
+            } else if (variantType == PaletteBlockPattern.LAYERED) {
+                variants = STONE_VARIANTS;
+                s = "layered_" + s;
+            } else if (variantType == PaletteBlockPattern.PILLAR) {
+                variants = STONE_VARIANTS;
+                s = s + "_pillar";
+            }
+            String finalS = s;
+            CREATE_MANAGERS.add(CaliberBlockManager.registerCreate("%s".formatted(s), properties, () -> ForgeRegistries.BLOCKS.getValue(new ResourceLocation("create", finalS)), variants));
+        }
+    }
 
     public static void printBlockCounts() {
-        System.out.println("Create blocks registered: " + blockCount);
+        System.out.println("Caliber Mod registered " + blockCount + " new Create blocks.");
     }
 
     private static <T extends Block> RegistryObject<T> registerBlockWithoutBlockItem(String name, Supplier<T> block) {